requirements.txt
pandas
numpy
flask
scikit-learn
seaborn
datetime
mysql-connector-python
sqlalchemy
mysql-client
-e .